Rename grid_with_result fil to gwr
