+++ tests/models/input/container/test_raw_grid.py
@@ -1,4 +1,4 @@
-from psdm_analysis.models.grid_with_results import GridWithResults
+from psdm_analysis.models.gwr import GridWithResults
 
 
 def test_raw_grid_container(gwr):
