+++ tests/models/test_gwr.py
@@ -2,7 +2,7 @@ from datetime import datetime
 
 import pytest
 
-from psdm_analysis.models.grid_with_results import GridWithResults
+from psdm_analysis.models.gwr import GridWithResults
 from psdm_analysis.models.result.grid.enhanced_node import EnhancedNodeResult
